@@ -14,64 +14,69 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^11.0.0",
-    "@angular/cdk": "^11.0.0",
-    "@angular/common": "^11.0.0",
-    "@angular/compiler": "^11.0.0",
-    "@angular/core": "^11.0.0",
-    "@angular/forms": "^11.0.0",
-    "@angular/material": "^11.0.0",
-    "@angular/platform-browser": "^11.0.0",
-    "@angular/platform-browser-dynamic": "^11.0.0",
-    "@angular/platform-server": "^11.0.0",
-    "@angular/router": "^11.0.0",
-    "@fortawesome/angular-fontawesome": "^0.8.0",
-    "@fortawesome/fontawesome-svg-core": "^1.2.32",
-    "@fortawesome/free-brands-svg-icons": "^5.15.1",
-    "@fortawesome/free-solid-svg-icons": "^5.15.1",
-    "ethers": "^5.0.19",
-    "express": "^4.15.2",
+    "@angular/animations": "^11.2.7",
+    "@angular/cdk": "^11.2.6",
+    "@angular/common": "^11.2.7",
+    "@angular/compiler": "^11.2.7",
+    "@angular/core": "^11.2.7",
+    "@angular/forms": "^11.2.7",
+    "@angular/material": "^11.2.6",
+    "@angular/platform-browser": "^11.2.7",
+    "@angular/platform-browser-dynamic": "^11.2.7",
+    "@angular/platform-server": "^11.2.7",
+    "@angular/router": "^11.2.7",
+    "@fortawesome/angular-fontawesome": "^0.8.2",
+    "@fortawesome/fontawesome-svg-core": "^1.2.35",
+    "@fortawesome/free-brands-svg-icons": "^5.15.3",
+    "@fortawesome/free-solid-svg-icons": "^5.15.3",
+    "@orvium/ux-components": "file:../orvium-ux/dist/orvium-ux-lib/orvium-ux-components-0.0.4.tgz",
+    "ethers": "^5.1.0",
+    "express": "^4.15.5",
     "html2canvas": "^1.0.0-rc.7",
+    "http-proxy-middleware": "^1.0.6",
     "ng-circle-progress": "^1.6.0",
     "ngx-gravatar": "^10.0.0",
-    "ngx-logger": "^4.1.9",
-    "ngx-sharebuttons": "^8.0.4",
+    "ngx-logger": "^4.2.1",
+    "ngx-sharebuttons": "^8.0.5",
     "ngx-smart-modal": "^7.4.1",
-    "ngx-spinner": "^10.0.1",
-    "rxjs": "^6.6.3",
+    "rxjs": "^6.6.7",
     "ts-md5": "^1.2.7",
     "tslib": "^2.0.3",
-    "zone.js": "~0.10.3"
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.1100.1",
-    "@angular-eslint/builder": "0.6.0-beta.0",
-    "@angular-eslint/eslint-plugin": "0.6.0-beta.0",
-    "@angular-eslint/eslint-plugin-template": "0.6.0-beta.0",
-    "@angular-eslint/schematics": "^0.6.0-beta.0",
-    "@angular-eslint/template-parser": "0.6.0-beta.0",
-    "@angular/cli": "^11.0.1",
-    "@angular/compiler-cli": "^11.0.0",
-    "@angular/language-service": "^11.0.0",
-    "@types/express": "^4.17.9",
-    "@types/jasmine": "~3.6.1",
+    "@angular-devkit/architect": "^0.1102.6",
+    "@angular-devkit/build-angular": "^0.1102.6",
+    "@angular-eslint/builder": "^2.0.2",
+    "@angular-eslint/eslint-plugin": "^2.0.2",
+    "@angular-eslint/eslint-plugin-template": "^2.0.2",
+    "@angular-eslint/schematics": "2.0.2",
+    "@angular-eslint/template-parser": "^2.0.2",
+    "@angular/cli": "^11.2.6",
+    "@angular/compiler-cli": "^11.2.7",
+    "@angular/language-service": "^11.2.7",
+    "@types/express": "^4.17.11",
+    "@types/jasmine": "~3.6.9",
     "@types/jasminewd2": "^2.0.8",
-    "@types/node": "^12.11.1",
-    "@typescript-eslint/eslint-plugin": "4.7.0",
-    "@typescript-eslint/parser": "4.7.0",
-    "eslint": "^7.13.0",
-    "jasmine-core": "~3.6.0",
+    "@types/node": "^12.20.6",
+    "@typescript-eslint/eslint-plugin": "4.20.0",
+    "@typescript-eslint/parser": "4.20.0",
+    "eslint": "^7.23.0",
+    "eslint-plugin-import": "2.22.1",
+    "eslint-plugin-jsdoc": "32.3.0",
+    "eslint-plugin-prefer-arrow": "1.2.3",
+    "jasmine-core": "~3.7.1",
     "jasmine-spec-reporter": "~6.0.0",
-    "karma": "~5.2.3",
+    "karma": "~6.3.2",
     "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage-istanbul-reporter": "~3.0.3",
+    "karma-coverage": "~2.0.3",
     "karma-jasmine": "~4.0.1",
     "karma-jasmine-html-reporter": "^1.5.4",
     "stylelint": "^13.7.2",
     "stylelint-config-standard": "^20.0.0",
     "stylelint-scss": "^3.18.0",
     "ts-node": "^8.8.2",
-    "typescript": "~4.0.5",
-    "webpack-bundle-analyzer": "^4.1.0"
+    "typescript": "~4.1.5",
+    "webpack-bundle-analyzer": "^4.4.0"
   }
 }
