@@ -2,13 +2,14 @@ import { Component, OnInit } from '@angular/core';
 import { NGXLogger } from 'ngx-logger';
 import { AbstractControl, FormBuilder, FormGroup, ValidationErrors, Validators } from '@angular/forms';
 import { OrviumService } from '../../services/orvium.service';
-import { Institution, Profile, USER_TYPE } from '../../model/orvium';
 import { ActivatedRoute, Router } from '@angular/router';
-import { catchError, debounceTime, map, tap } from 'rxjs/operators';
-import { CustomValidators } from '../../shared/CustomValidators';
+import { catchError, map } from 'rxjs/operators';
+import { AppCustomValidators } from '../../shared/AppCustomValidators';
 import errorMessagesJSON from '../../../assets/messages.json';
 import { Md5 } from 'ts-md5';
 import { Observable, of } from 'rxjs';
+import { ProfileService } from '../profile.service';
+import { USER_TYPE, UserPrivateDTO } from 'src/app/model/api';
 
 @Component({
   selector: 'app-first-login',
@@ -16,56 +17,49 @@ import { Observable, of } from 'rxjs';
   styleUrls: ['./onboarding.component.scss']
 })
 export class OnboardingComponent implements OnInit {
+  profile: UserPrivateDTO;
   userTypeFormGroup: FormGroup;
   userProfileFormGroup: FormGroup;
-  userInstitutionFormGroup: FormGroup;
   done = false;
-  gravatar: string;
-  institution: Institution;
+  gravatar = '';
   errorMessages = errorMessagesJSON;
   USER_TYPE = USER_TYPE;
+  canContinue = false;
 
   constructor(private logger: NGXLogger,
               private formBuilder: FormBuilder,
               private orviumService: OrviumService,
+              private profileService: ProfileService,
               private router: Router,
               private route: ActivatedRoute) {
-  }
-
-  ngOnInit(): void {
+    this.profile = this.route.snapshot.data.profile;
 
     this.userTypeFormGroup = this.formBuilder.group({
       userType: ['', Validators.required]
     });
 
     this.userProfileFormGroup = this.formBuilder.group({
-      email: [null, [Validators.required, CustomValidators.validateEmail], [this.validateDomain.bind(this)]],
+      email: [null, [Validators.required, AppCustomValidators.validateEmail], [this.validateDomain.bind(this)]],
       firstName: [null, Validators.required],
-      lastName: [null, Validators.required]
+      lastName: [null, Validators.required],
+      acceptedTC: [false, Validators.required]
     });
+  }
 
-    this.userInstitutionFormGroup = this.formBuilder.group({});
-
-    this.userProfileFormGroup.controls.email.valueChanges
-      .pipe(
-        debounceTime(300),
-        tap(() => this.onBlurInstitutionalEmail())
-      ).subscribe();
-
-    this.populateFormOnLoad(this.route.snapshot.data.profile);
-
+  ngOnInit(): void {
+    this.populateFormOnLoad(this.profile);
   }
 
 
   validateDomain(control: AbstractControl): Promise<ValidationErrors | null> | Observable<ValidationErrors | null> {
     const domain = control.value.replace(/.*@/, '');
     return this.orviumService.getDomain(domain).pipe(
+      catchError(() => of(null)),
       map(invalidDomain => invalidDomain ? { invalidDomain: true } : null),
-      catchError(() => of(null))
     );
   }
 
-  populateFormOnLoad(profile: Profile): void {
+  populateFormOnLoad(profile: UserPrivateDTO): void {
 
     this.userTypeFormGroup.get('userType')?.setValue(profile.userType);
 
@@ -78,30 +72,28 @@ export class OnboardingComponent implements OnInit {
     this.gravatar = Md5.hashStr(profile.email || '').toString();
   }
 
-  onBlurInstitutionalEmail(): void {
-    const emailField = this.userProfileFormGroup.get('email');
-    if (emailField && emailField.valid) {
-      const emailDomain = emailField.value.replace(/.*@/, '');
-
-      this.orviumService.getInstitutionByDomain(emailDomain)
-        .subscribe(institution => {
-          console.log('the institution', institution);
-          this.institution = institution;
-        });
-    }
-  }
-
   save(): void {
-    let profile = new Profile();
+    let profile = new UserPrivateDTO();
     profile.isOnboarded = true;
     profile = Object.assign(profile, this.userTypeFormGroup.value);
     profile = Object.assign(profile, this.userProfileFormGroup.value);
     profile.isOnboarded = true;
-
-    this.orviumService.updateProfile(profile).subscribe(response => {
+    this.profileService.updateProfile(profile).subscribe(response => {
       this.done = true;
       this.logger.debug(response);
       this.router.navigate(['/profile']);
     });
+  }
+
+  getErrorMessage(errors: ValidationErrors | null): string | null {
+    if (!errors) {
+      return null;
+    }
+    let keys = Object.keys(errors);
+    if (keys.includes('required')) {
+      return 'You must enter a value';
+    }
+
+    return errors[keys[0]];
   }
 }
