@@ -3,15 +3,13 @@ import { AppModule } from './app.module';
 import { AppComponent } from './app.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
-
 @NgModule({
   declarations: [],
   imports: [
     AppModule,
     BrowserAnimationsModule,
   ],
-  providers: [
-  ],
+  providers: [],
   bootstrap: [AppComponent]
 })
 export class AppBrowserModule {
