@@ -4,7 +4,7 @@ import { MatIconRegistry } from '@angular/material/icon';
 import { OrviumService } from './services/orvium.service';
 import { Router } from '@angular/router';
 import { environment } from '../environments/environment';
-import { DOCUMENT, LocationStrategy } from '@angular/common';
+import { DOCUMENT } from '@angular/common';
 import { BlockchainService } from './blockchain/blockchain.service';
 import { FaIconLibrary } from '@fortawesome/angular-fontawesome';
 import { faFileWord } from '@fortawesome/free-solid-svg-icons/faFileWord';
@@ -18,9 +18,16 @@ import { faLeaf } from '@fortawesome/free-solid-svg-icons/faLeaf';
 import { faTwitter } from '@fortawesome/free-brands-svg-icons/faTwitter';
 import { faFacebookF } from '@fortawesome/free-brands-svg-icons/faFacebookF';
 import { faLinkedin } from '@fortawesome/free-brands-svg-icons/faLinkedin';
-import { Profile } from './model/orvium';
+import { faCreativeCommons } from '@fortawesome/free-brands-svg-icons/faCreativeCommons';
+import { faCreativeCommonsZero } from '@fortawesome/free-brands-svg-icons/faCreativeCommonsZero';
+import { faCreativeCommonsBy } from '@fortawesome/free-brands-svg-icons/faCreativeCommonsBy';
+import { faCreativeCommonsNd } from '@fortawesome/free-brands-svg-icons/faCreativeCommonsNd';
 import { faOrcid } from '@fortawesome/free-brands-svg-icons/faOrcid';
+import { faGithub } from '@fortawesome/free-brands-svg-icons/faGithub';
 import { ThemeService } from './theme/theme.service';
+import { faEthereum } from '@fortawesome/free-brands-svg-icons/faEthereum';
+import { ProfileService } from './profile/profile.service';
+import { UserPrivateDTO } from './model/api';
 
 @Component({
   selector: 'app-root',
@@ -28,12 +35,13 @@ import { ThemeService } from './theme/theme.service';
   styleUrls: ['./app.component.scss']
 })
 export class AppComponent implements OnInit {
-  public profile: Profile;
+  public profile?: UserPrivateDTO;
   public theme = 'light-theme';
 
   constructor(private iconRegistry: MatIconRegistry,
               private domSanitizer: DomSanitizer,
               private orviumService: OrviumService,
+              private profileService: ProfileService,
               private blockchainService: BlockchainService,
               private router: Router,
               private metaService: Meta,
@@ -46,8 +54,6 @@ export class AppComponent implements OnInit {
       this.domSanitizer.bypassSecurityTrustResourceUrl('https://assets.orvium.io/logo/logo.svg'));
     this.iconRegistry.addSvgIcon('orviumIcon',
       this.domSanitizer.bypassSecurityTrustResourceUrl('https://assets.orvium.io/logo/orviumIcon.svg'));
-    this.iconRegistry.addSvgIcon('ethereum',
-      this.domSanitizer.bypassSecurityTrustResourceUrl('https://assets.orvium.io/logo/ethereum-logo.svg'));
   }
 
   ngOnInit(): void {
@@ -55,11 +61,10 @@ export class AppComponent implements OnInit {
       this.metaService.addTag({ name: 'robots', content: 'noindex' });
     }
 
-    this.orviumService.getProfile().subscribe(profileRefreshed => {
-      if (profileRefreshed) {
-        this.profile = profileRefreshed;
-      }
-    });
+    const profile = this.profileService.profile.getValue();
+    if (profile) {
+      this.profile = profile;
+    }
 
     this.blockchainService.initNetworks();
 
@@ -76,10 +81,16 @@ export class AppComponent implements OnInit {
       faTwitter,
       faFacebookF,
       faLinkedin,
-      faOrcid
+      faOrcid,
+      faEthereum,
+      faCreativeCommons,
+      faCreativeCommonsZero,
+      faCreativeCommonsBy,
+      faCreativeCommonsNd,
+      faGithub
     );
 
-    this.themeService.getTheme().subscribe( theme => this.theme = theme);
+    this.themeService.getTheme().subscribe(theme => this.theme = theme);
   }
 
   resetPosition(): void {
