@@ -1,14 +1,14 @@
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
-import { Network } from '../model/orvium';
 import { environment } from '../../environments/environment';
 import { HttpClient } from '@angular/common/http';
+import { BlockchainNetworkDTO } from '../model/api';
 
 @Injectable({
   providedIn: 'root'
 })
 export class BlockchainService {
-  networks: Network[];
+  networks: BlockchainNetworkDTO[] = [];
 
   constructor(private http: HttpClient) {
   }
@@ -17,11 +17,11 @@ export class BlockchainService {
     this.getAllNetworkConfigs().subscribe(networks => this.networks = networks);
   }
 
-  getAllNetworkConfigs(): Observable<Network[]> {
-    return this.http.get<Network[]>(`${environment.apiEndpoint}/blockchain`);
+  getAllNetworkConfigs(): Observable<BlockchainNetworkDTO[]> {
+    return this.http.get<BlockchainNetworkDTO[]>(`${environment.apiEndpoint}/blockchain`);
   }
 
-  getNetworkByName(networkName: string): Network | null {
+  getNetworkByName(networkName: string): BlockchainNetworkDTO | null {
     for (const network of this.networks) {
       if (network.name === networkName) {
         return network;
@@ -30,7 +30,7 @@ export class BlockchainService {
     return null;
   }
 
-  getNetworkConfig(networkId: number): Network | undefined {
+  getNetworkConfig(networkId: number): BlockchainNetworkDTO | undefined {
     for (const network of this.networks) {
       if (network.networkId === networkId) {
         return network;
