@@ -1,13 +1,15 @@
 import { Component, Inject, OnInit } from '@angular/core';
 import { Meta, Title } from '@angular/platform-browser';
 import { DOCUMENT } from '@angular/common';
-import { Community, Deposit, Profile } from '../../model/orvium';
 import { OrviumService } from '../../services/orvium.service';
 import { ActivatedRoute, Router } from '@angular/router';
-import { MatSnackBar } from '@angular/material/snack-bar';
 import { NgxSmartModalService } from 'ngx-smart-modal';
 import { FormControl, FormGroup } from '@angular/forms';
 import { ShareService } from 'ngx-sharebuttons';
+import { ProfileService } from '../../profile/profile.service';
+import { AppSnackBarService } from '../../services/app-snack-bar.service';
+import { CommunityDTO, CreateDepositDTO, DepositDTO, UserPrivateDTO, UserSummaryDTO } from '../../model/api';
+import { CommunityService } from '../community.service';
 
 @Component({
   selector: 'app-community-view',
@@ -15,11 +17,14 @@ import { ShareService } from 'ngx-sharebuttons';
   styleUrls: ['./community-view.component.scss']
 })
 export class CommunityViewComponent implements OnInit {
-  community: Community;
-  profile: Profile;
+  community!: CommunityDTO;
+  profile?: UserPrivateDTO;
   isContributor = false;
-  deposits: Deposit[];
-  newPublicationForm: FormGroup;
+  deposits: DepositDTO[] = [];
+  moderators: UserSummaryDTO[] = [];
+  newPublicationForm = new FormGroup({
+    title: new FormControl(''),
+  });
   public isModerator = false;
 
   constructor(
@@ -27,8 +32,10 @@ export class CommunityViewComponent implements OnInit {
     private titleService: Title,
     private metaService: Meta,
     private orviumService: OrviumService,
+    private profileService: ProfileService,
+    private communityService: CommunityService,
     private route: ActivatedRoute,
-    private snackBar: MatSnackBar,
+    private snackBar: AppSnackBarService,
     public router: Router,
     public ngxSmartModalService: NgxSmartModalService,
     @Inject(DOCUMENT) private dom: Document
@@ -36,23 +43,13 @@ export class CommunityViewComponent implements OnInit {
   }
 
   ngOnInit(): void {
-    this.route.data.pipe().subscribe(async data => {
+    this.route.data.subscribe(async data => {
       this.community = data.community;
-    });
-    this.orviumService.getProfile().subscribe(profile => {
-      if (profile) {
-        this.profile = profile;
-        this.isModerator = this.profile.roles.includes('moderator:' + this.community._id);
+      if (!this.community) {
+        throw new Error('Community not found');
       }
+      this.init();
     });
-    this.orviumService.getCommunityDeposits(this.community._id).subscribe(deposits => {
-      this.deposits = deposits;
-    });
-    this.newPublicationForm = new FormGroup({
-      title: new FormControl(''),
-    });
-    this.setIsContributor();
-    this.setSEOTags();
   }
 
   public setSEOTags(): void {
@@ -60,12 +57,12 @@ export class CommunityViewComponent implements OnInit {
     this.metaService.addTags([
       {
         name: 'description',
-        content: this.community.description
+        content: this.community.description || this.community.name
       },
       { name: 'og:title', content: this.community.name + ' | Orvium' },
       {
         name: 'og:description',
-        content: this.community.description,
+        content: this.community.description || this.community.name
       },
       { name: 'og:url', content: this.dom.URL },
       { name: 'og:site_name', content: this.community.name }
@@ -73,16 +70,9 @@ export class CommunityViewComponent implements OnInit {
   }
 
   public joinCommunity(): void {
-    if (!this.profile) {
-      this.snackBar.open('You need to login first',
-        'Dismiss', { panelClass: ['info-snackbar'] });
-      return;
-    }
-
     this.orviumService.joinCommunity(this.community._id).subscribe(community => {
-      this.snackBar.open('Congratulations, you have joined ' + this.community.name + ' community',
-        'Dismiss', { panelClass: ['ok-snackbar'] });
-      this.orviumService.getProfileFromApi().subscribe(profile => {
+      this.snackBar.info('Congratulations, you have joined ' + this.community.name + ' community');
+      this.profileService.getProfileFromApi().subscribe(profile => {
         this.setIsContributor();
       });
       this.community = community;
@@ -91,21 +81,23 @@ export class CommunityViewComponent implements OnInit {
   }
 
   public setIsContributor(): void {
-    if (this.profile && this.profile.communities?.includes(this.community._id)) {
-      this.isContributor = true;
+    if (this.profile) {
+      for (const community of this.profile.communities) {
+        if (community._id === this.community._id) {
+          this.isContributor = true;
+        }
+      }
     }
   }
 
   public uploadDeposit(): void {
     if (!this.profile) {
-      this.snackBar.open('You need to login first',
-        'Dismiss', { panelClass: ['info-snackbar'] });
+      this.snackBar.info('You need to login first');
       return;
     }
 
     if (!this.profile?.isOnboarded) {
-      this.snackBar.open('Complete your profile first to upload a publication',
-        'Dismiss', { panelClass: ['info-snackbar'] });
+      this.snackBar.info('Complete your profile first to upload a publication');
     } else {
       this.ngxSmartModalService.open('uploadDepositModal');
     }
@@ -113,22 +105,42 @@ export class CommunityViewComponent implements OnInit {
 
   public createDeposit(): void {
     if (this.newPublicationForm.get('title')?.value !== '' && this.newPublicationForm.get('title')?.value !== null) {
-      const deposit = new Deposit();
+      const deposit = new CreateDepositDTO();
       deposit.title = this.newPublicationForm.get('title')?.value;
-      deposit.community = this.community;
+      deposit.community = this.community._id;
       this.orviumService.createDeposit(deposit).subscribe(response => {
         this.newPublicationForm.reset();
         this.router.navigate(['deposits', response._id, 'edit']);
       });
     } else {
       this.newPublicationForm.reset();
-      this.snackBar.open('Please, enter a title for your publication', 'Dismiss', { panelClass: ['error-snackbar'] });
+      this.snackBar.error('Please, enter a title for your publication');
     }
-    this.close();
-  }
-
-  close(): void {
     this.newPublicationForm.reset();
     this.ngxSmartModalService.close('uploadDepositModal');
+  }
+
+  public callForPapers(): void {
+    this.ngxSmartModalService.open('callForPapersModal');
+  }
+
+  private init(): void {
+
+    this.profileService.getProfile().subscribe(profile => {
+      if (profile) {
+        this.profile = profile;
+        this.isModerator = this.communityService.canModerateCommunity(this.community);
+
+      }
+    });
+    this.orviumService.getCommunityDeposits(this.community._id).subscribe(deposits => {
+      this.deposits = deposits;
+    });
+    if (this.community.moderators) {
+      this.moderators = this.community.moderators;
+    }
+
+    this.setIsContributor();
+    this.setSEOTags();
   }
 }
