@@ -1,22 +1,24 @@
-import { Component, OnInit } from '@angular/core';
+import { Component } from '@angular/core';
+import { Feedback } from '../shared/feedback/entity/feedback';
 import { OrviumService } from '../services/orvium.service';
-import { MatSnackBar } from '@angular/material/snack-bar';
+import { AppSnackBarService } from '../services/app-snack-bar.service';
 
 @Component({
   selector: 'app-dashboard',
   templateUrl: './dashboard.component.html',
   styleUrls: ['./dashboard.component.scss']
 })
-export class DashboardComponent implements OnInit {
+export class DashboardComponent {
 
   constructor(private orviumService: OrviumService,
-              private snackBar: MatSnackBar) {
-  }
-
-  ngOnInit(): void {
+              private snackBar: AppSnackBarService) {
   }
 
   onSend(event: object): void {
+    const feedback = event as Feedback;
+    this.orviumService.createFeedback(feedback).subscribe(data => {
+      this.snackBar.info('Thank you for your feedback!');
+    });
   }
 
 }
