@@ -5,13 +5,11 @@ import { MatSnackBarModule } from '@angular/material/snack-bar';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { RouterTestingModule } from '@angular/router/testing';
 import { ActivatedRoute } from '@angular/router';
-import { ACCESS_RIGHT, Deposit, DEPOSIT_STATUS, OrviumFile, Profile, PUBLICATION_TYPE, REVIEW_TYPE } from '../../model/orvium';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { NgxSmartModalModule } from 'ngx-smart-modal';
 import { HarnessLoader } from '@angular/cdk/testing';
 import { TestbedHarnessEnvironment } from '@angular/cdk/testing/testbed';
 import { MatAutocompleteModule } from '@angular/material/autocomplete';
-import { MatButtonHarness } from '@angular/material/button/testing';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatCardModule } from '@angular/material/card';
@@ -20,13 +18,11 @@ import { MatTabsModule } from '@angular/material/tabs';
 import { MatStepperModule } from '@angular/material/stepper';
 import { MatIconModule } from '@angular/material/icon';
 import { MatNativeDateModule } from '@angular/material/core';
-import { NgxSpinnerModule } from 'ngx-spinner';
 import { MatChipsModule } from '@angular/material/chips';
 import { MatTableModule } from '@angular/material/table';
 import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
 import { MatTooltipModule } from '@angular/material/tooltip';
 import { GravatarModule } from 'ngx-gravatar';
-import { FileuploadComponent } from '../../shared/fileupload/fileupload.component';
 import { MatInputModule } from '@angular/material/input';
 import { MatSelectModule } from '@angular/material/select';
 import { DepositsReviewsTableComponent } from '../deposits-reviews-table/deposits-reviews-table.component';
@@ -35,52 +31,56 @@ import { DisciplinesService } from '../../services/disciplines.service';
 import { MatButtonModule } from '@angular/material/button';
 import { LoggerTestingModule } from 'ngx-logger/testing';
 import { FontAwesomeTestingModule } from '@fortawesome/angular-fontawesome/testing';
-import { AccessDeniedComponent } from '../../shared/access-denied/access-denied.component';
-import { profileTest } from '../../shared/test-data';
+import { depositDraft, fileTest, profilePrivateTest } from '../../shared/test-data';
+import { DepositsService } from '../deposits.service';
+import { ProfileService } from '../../profile/profile.service';
+import { BehaviorSubject } from 'rxjs';
+import { Component, Input } from '@angular/core';
+import { DepositDTO, FileMetadata, UserPrivateDTO } from '../../model/api';
+import { of } from 'rxjs';
+
+@Component({ selector: 'app-fileupload', template: '' })
+class FileUploadStubComponent {
+  @Input() url?: string;
+  @Input() accept?: string;
+  @Input() maxFileSize?: number;
+  @Input() fileLimit?: number;
+  @Input() uploadedFileCount?: number;
+  @Input() isPublication?: boolean;
+}
+
+@Component({ selector: 'app-file-list', template: '' })
+class FileListStubComponent {
+  @Input() files: FileMetadata[] = [];
+  @Input() baseHref?: string;
+  @Input() readonly = true;
+}
 
 describe('DepositDetailsComponent', () => {
   let component: DepositDetailsComponent;
   let fixture: ComponentFixture<DepositDetailsComponent>;
   let loader: HarnessLoader;
 
-  const profile: Profile = { ...profileTest, ...{ isOnboarded: true } };
-
-  const file: OrviumFile = {
-    filename: 'file',
-    contentType: 'image/png',
-    keccak256: '0x1111111111111111111111111111111111111111111111111111111111111111',
-    contentLength: 21500,
-  };
-
-  const deposit: Deposit = {
-    _id: '123412341234',
-    owner: 'theowner',
-    title: 'the title',
-    abstract: 'the abstract',
-    publicationType: PUBLICATION_TYPE.book,
-    accessRight: ACCESS_RIGHT.CC0,
-    publicationDate: '1968-11-16T00:00:00',
-    status: DEPOSIT_STATUS.draft,
-    peerReviews: [],
-    reviewType: REVIEW_TYPE.openReview,
+  const profile: UserPrivateDTO = { ...profilePrivateTest(), ...{ isOnboarded: true, emailConfirmed: true } };
+  const deposit: DepositDTO = {
+    ...depositDraft(),
+    disciplines: ['Acoustics', 'Computing'],
     authors: [{ name: 'John', surname: 'Doe' }, { name: 'William', surname: 'Wallace' }],
     keywords: ['science', 'cloud'],
-    keccak256: '0x1111111111111111111111111111111111111111111111111111111111111111',
-    files: [],
-    doi: '10.1000/182',
-    url: '',
-    pdfUrl: '',
-    gravatar: '11111111111111111111111111111111',
-    disciplines: ['Abnormal psychology', 'Acoustics'],
-    references: [],
-    publicationFile: file,
+    publicationFile: fileTest()
   };
+  const file = fileTest;
 
   const routeSnapshot = { snapshot: { data: { deposit } } };
 
   beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
-      declarations: [DepositDetailsComponent, FileuploadComponent, DepositsReviewsTableComponent, AccessDeniedComponent],
+      declarations: [
+        DepositDetailsComponent,
+        DepositsReviewsTableComponent,
+        FileUploadStubComponent,
+        FileListStubComponent
+      ],
       imports: [RouterTestingModule, MatSnackBarModule,
         FormsModule, ReactiveFormsModule, HttpClientModule,
         HttpClientTestingModule, MatAutocompleteModule,
@@ -91,7 +91,6 @@ describe('DepositDetailsComponent', () => {
         MatSelectModule,
         MatDatepickerModule,
         MatNativeDateModule,
-        NgxSpinnerModule,
         MatChipsModule,
         MatTableModule,
         FontAwesomeModule,
@@ -114,8 +113,11 @@ describe('DepositDetailsComponent', () => {
     loader = TestbedHarnessEnvironment.loader(fixture);
     component = fixture.componentInstance;
     const disciplinesService = fixture.debugElement.injector.get(DisciplinesService);
-    spyOn(disciplinesService, 'getDisciplines').and.returnValue([]);
-
+    const depositService = fixture.debugElement.injector.get(DepositsService);
+    spyOn(disciplinesService, 'getDisciplines').and.returnValue(of([]));
+    spyOn(depositService, 'canUpdateDeposit').and.returnValue(true);
+    const profileService = fixture.debugElement.injector.get(ProfileService);
+    profileService.profile = new BehaviorSubject<UserPrivateDTO | undefined>(profile);
     fixture.detectChanges();
   }));
 
@@ -130,17 +132,17 @@ describe('DepositDetailsComponent', () => {
   it('should remove a discipline', () => {
     component.canManageDeposit = true;
     component.canUpdateDeposit = true;
-    expect(component.deposit?.disciplines?.length).toBe(2);
+    expect(component.deposit?.disciplines.length).toBe(2);
     component.removeDiscipline('Acoustics');
-    expect(component.deposit?.disciplines?.length).toBe(1);
+    expect(component.deposit?.disciplines.length).toBe(1);
   });
 
   it('should remove a keyword', () => {
     component.canManageDeposit = true;
     component.canUpdateDeposit = true;
-    expect(component.deposit?.keywords?.length).toBe(2);
+    expect(component.deposit?.keywords.length).toBe(2);
     component.removeKeyword('science');
-    expect(component.deposit?.keywords?.length).toBe(1);
+    expect(component.deposit?.keywords.length).toBe(1);
   });
 
   it('should remove an author', () => {
@@ -152,8 +154,7 @@ describe('DepositDetailsComponent', () => {
   });
 
   it('should submit to preprint', async () => {
-    const profile: Profile = { ...profileTest, ...{ userId: deposit.owner } };
-    component.profile = profile;
+    component.profile = { ...profilePrivateTest(), userId: deposit.owner };
     component.refreshDeposit(deposit);
     component.depositForm.markAsPristine();
     expect(component.canBeSentToReview()).toBe(true);
