@@ -1,7 +1,7 @@
 import { Component, Input, OnInit } from '@angular/core';
-import { PeerReview, Profile, REVIEW_DECISION, REVIEW_STATUS } from '../../model/orvium';
-import { OrviumService } from '../../services/orvium.service';
 import { environment } from 'src/environments/environment';
+import { ProfileService } from '../../profile/profile.service';
+import { REVIEW_DECISION, REVIEW_STATUS, ReviewDTO, UserPrivateDTO } from 'src/app/model/api';
 
 @Component({
   selector: 'app-deposits-reviews-table',
@@ -9,22 +9,24 @@ import { environment } from 'src/environments/environment';
   styleUrls: ['./deposits-reviews-table.component.scss']
 })
 export class DepositsReviewsTableComponent implements OnInit {
-  @Input() reviews: PeerReview[];
+  @Input() reviews: ReviewDTO[] = [];
   @Input() basic = false;
 
   reviewColumnsToDisplay = ['author', 'creationDate', 'publicationDate', 'status', 'decision', 'file', 'actions'];
-  reviewColumnsToDisplayBasic = ['author', 'status'];
+  reviewColumnsToDisplayBasic = ['author', 'status', 'actions'];
   REVIEW_STATUS = REVIEW_STATUS;
   REVIEW_DECISION = REVIEW_DECISION;
   environment = environment;
   selectedReviewer: unknown;
-  profile: Profile;
+  profile?: UserPrivateDTO;
 
-  constructor(private orviumService: OrviumService) {
+  constructor(
+    private profileService: ProfileService,
+  ) {
   }
 
   ngOnInit(): void {
-    this.orviumService.getProfile().subscribe(profile => {
+    this.profileService.getProfile().subscribe(profile => {
       if (profile) {
         this.profile = profile;
       }
