@@ -1,251 +1,26 @@
-import { TransactionReceipt, TransactionResponse } from '@ethersproject/providers';
-
-export enum PUBLICATION_TYPE {
-  book = 'book',
-  bookSection = 'book section',
-  conferencePaper = 'conference paper',
-  article = 'article',
-  patent = 'patent',
-  preprint = 'preprint',
-  report = 'report',
-  softwareDocumentation = 'software documentation',
-  thesis = 'thesis',
-  technicalNote = 'technical note',
-  workingPaper = 'working paper',
-  policyReport = 'policy report',
-  registeredReport = 'registered report',
-  proposal = 'proposal',
-  reviewArticle = 'review article',
-  video = 'video',
-  other = 'other'
-}
-
-export enum ACCESS_RIGHT {
-  CCBY = 'cc by',
-  CCBYND = 'cc by-nd',
-  CC0 = 'cc0'
-}
-
-export enum REVIEW_TYPE {
-  openReview = 'open review',
-  singleBlind = 'single blind',
-  doubleBlind = 'double blind'
-}
-
-export enum DEPOSIT_STATUS {
-  draft = 'draft',
-  pendingApproval = 'pending approval',
-  preprint = 'preprint',
-  inReview = 'in review',
-  published = 'published'
-}
-
-export enum INVITE_STATUS {
-  pending = 'draft',
-  accepted = 'accepted',
-  rejected = 'rejected',
-}
-
-export enum REVIEW_STATUS {
-  draft = 'draft',
-  published = 'published'
-}
-
-export enum CREDIT_TYPE {
-  methodology = 'methodology',
-  conceptualization = 'conceptualization',
-  software = 'software',
-  validation = 'validation',
-  formalAnalysis = 'formal analysis',
-  investigation = 'investigation',
-  resources = 'resources',
-  dataCuration = 'data curation',
-  writingOriginalDraft = 'writing original draft',
-  writingReviewEditing = 'writing review and editing',
-  visualization = 'visualization',
-  supervision = 'supervision',
-  projectAdministration = 'project administration',
-  fundingAcquisition = 'funding acquisition',
-}
-
-export enum REVIEW_DECISION {
-  accepted = 'accepted',
-  minorRevision = 'minor revision',
-  mayorRevision = 'mayor revision'
-}
-
-export enum USER_TYPE {
-  student = 'student',
-  medical = 'medical',
-  business = 'business',
-  researcher = 'researcher',
-  citizen = 'citizen'
-}
-
-export class Deposit {
-  _id: string;
-  owner: string;
-  title: string;
-  abstract: string;
-  publicationType: PUBLICATION_TYPE;
-  accessRight: ACCESS_RIGHT;
-  submissionDate?: string;
-  publicationDate?: string;
-  status: DEPOSIT_STATUS;
-  peerReviews: PeerReview[];
-  reviewType: REVIEW_TYPE;
-  authors: Author[];
-  transactions?: Record<string, TransactionResponse | TransactionReceipt>;
-  publicationFile?: OrviumFile;
-  files: OrviumFile[];
-  gravatar?: string;
-  keywords: string[];
-  keccak256?: string;
-  doi?: string;
-  url?: string;
-  pdfUrl?: string;
-  disciplines?: string[];
-  references?: Reference[];
-  createdOn?: Date;
-  version?: number;
-  community?: Community;
-  html?: string;
-}
+import { COMMUNITY_TYPE, CREDIT_TYPE, DEPOSIT_STATUS, DepositDTO, PUBLICATION_TYPE, REVIEW_DECISION } from './api';
 
 export class DepositsQuery {
-  deposits: Deposit[];
-  count: number;
+  deposits!: DepositDTO[];
+  count!: number;
 }
 
-export class Reference {
-  reference: string;
-  url: string;
+export class TopDisciplinesQuery {
+  _id!: string;
+  count!: number;
 }
 
 export class Citation {
-  apa: string;
-}
-
-export class Author {
-  name: string;
-  surname: string;
-  email?: string;
-  orcid?: string;
-  credit?: CREDIT_TYPE[];
-}
-
-export class OrviumFile {
-  filename: string;
-  contentType: string;
-  keccak256: string;
-  contentLength: number;
-}
-
-export class Profile {
-  starredDeposits?: string[];
-  userId: string;
-  isReviewer: boolean;
-  isOnboarded: boolean;
-  firstName?: string;
-  lastName?: string;
-  institution?: string;
-  email?: string;
-  emailConfirmed: boolean;
-  orcid?: string;
-  userType: string;
-  disciplines: string[];
-  aboutMe?: string;
-  gravatar?: string;
-  blog?: string;
-  role?: string;
-  roles: string[];
-  linkedin?: string;
-  percentageComplete: number;
-  inviteToken?: string;
-  communities?: string[];
-  simultaneousReviews?: number;
-}
-
-export class Institution {
-  name: string;
-  domain: string;
-  country: string;
-  city: string;
-  synonym: string;
-}
-
-export class Discipline {
-  name: string;
-}
-
-export class Domain {
-  emailDomain: string;
-}
-
-
-export class PeerReview {
-  _id: string;
-  owner: string;
-  author?: string;
-  comments?: string;
-  decision?: REVIEW_DECISION;
-  file?: OrviumFile;
-  transactions?: Record<string, TransactionResponse | TransactionReceipt>;
-  url?: string;
-  fileUrl?: string;
-  status: REVIEW_STATUS;
-  gravatar?: string;
-  reward?: number;
-  revealReviewerIdentity?: boolean;
-  deposit: Deposit;
-  creationDate: string;
-  publicationDate: string;
-  wasInvited?: boolean;
-}
-
-export class Network {
-  name: string;
-  networkId: number;
-  displayName: string;
-  tokenAddress: string;
-  escrowAddress: string;
-  appAddress: string;
-  explorerUrl: string;
-}
-
-export class AppNotification {
-  _id: string;
-  title: string;
-  body: string;
-  icon: string;
-  date: Date;
-  isRead: boolean;
-  action: string;
-}
-
-export class Invite {
-  _id: string;
-  inviteType: string;
-  status: string;
-  deadline: Date;
-  sender: string;
-  addressee: string;
-  createdOn: Date;
-  data: unknown;
+  apa!: string;
 }
 
 export const PUBLICATION_TYPE_LOV = [
-  // { value: PUBLICATION_TYPE.book, viewValue: 'Book' },
-  // { value: PUBLICATION_TYPE.bookSection, viewValue: 'Book section' },
   { value: PUBLICATION_TYPE.conferencePaper, viewValue: 'Conference paper' },
   { value: PUBLICATION_TYPE.article, viewValue: 'Research article' },
-  // { value: PUBLICATION_TYPE.patent, viewValue: 'Patent' },
   { value: PUBLICATION_TYPE.preprint, viewValue: 'Preprint' },
   { value: PUBLICATION_TYPE.report, viewValue: 'Research report' },
   { value: PUBLICATION_TYPE.softwareDocumentation, viewValue: 'Software documentation' },
-  // { value: PUBLICATION_TYPE.thesis, viewValue: 'Thesis' },
   { value: PUBLICATION_TYPE.technicalNote, viewValue: 'Method paper' },
-  // { value: PUBLICATION_TYPE.workingPaper, viewValue: 'Working paper' },
   { value: PUBLICATION_TYPE.policyReport, viewValue: 'Policy report' },
   { value: PUBLICATION_TYPE.registeredReport, viewValue: 'Registered report' },
   { value: PUBLICATION_TYPE.proposal, viewValue: 'Research proposal' },
@@ -265,6 +40,11 @@ export const DEPOSIT_STATUS_LOV = [
   { value: DEPOSIT_STATUS.pendingApproval, viewValue: 'Pending approval' },
   { value: DEPOSIT_STATUS.preprint, viewValue: 'Preprint' },
   { value: DEPOSIT_STATUS.inReview, viewValue: 'In review' },
+  { value: DEPOSIT_STATUS.published, viewValue: 'Published' },
+];
+
+export const DEPOSIT_STATUS_PUBLIC_LOV = [
+  { value: DEPOSIT_STATUS.preprint, viewValue: 'Preprint' },
   { value: DEPOSIT_STATUS.published, viewValue: 'Published' },
 ];
 
@@ -402,25 +182,6 @@ export const REVIEW_DECISION_LOV: ReviewDecisionLov[] = [
 // Community
 // **************************
 
-export class Community {
-  _id: string;
-  name: string;
-  description: string;
-  country: string;
-  twitterURL: string;
-  facebookURL: string;
-  websiteURL: string;
-  users: CommunityUser[];
-  logoURL: string;
-  guidelinesURL: string;
-  acknowledgement: string;
-}
-
-export class CommunityUser {
-  userId: string;
-  role: string;
-}
-
 export enum COMMUNITY_ROLES {
   contributor = 'contributor',
 }
@@ -432,3 +193,13 @@ export const COMMUNITY_ROLES_LOV = [
   },
 ];
 
+export const COMMUNITY_TYPE_LOV = [
+  {
+    value: COMMUNITY_TYPE.university,
+    viewValue: 'University',
+  },
+  {
+    value: COMMUNITY_TYPE.business,
+    viewValue: 'Business',
+  },
+];
