@@ -1,22 +1,22 @@
 import { ApplicationRef, Inject, Injectable, PLATFORM_ID } from '@angular/core';
 import { BehaviorSubject, concat, Observable, timer } from 'rxjs';
-import { AppNotification } from '../model/orvium';
 import { environment } from '../../environments/environment';
 import { HttpClient } from '@angular/common/http';
 import { first, retry, switchMap, tap } from 'rxjs/operators';
-import { OrviumService } from '../services/orvium.service';
 import { isPlatformBrowser } from '@angular/common';
+import { ProfileService } from '../profile/profile.service';
+import { AppNotificationDTO } from '../model/api';
 
 @Injectable({
   providedIn: 'root'
 })
 export class NotificationService {
-  private notifications = new BehaviorSubject<AppNotification[]>([]);
+  private notifications = new BehaviorSubject<AppNotificationDTO[]>([]);
 
   constructor(
     private http: HttpClient,
     private appRef: ApplicationRef,
-    private orviumService: OrviumService,
+    private profileService: ProfileService,
     @Inject(PLATFORM_ID) private platformId: string
   ) {
 
@@ -25,7 +25,7 @@ export class NotificationService {
     }
 
     const appIsStable = appRef.isStable.pipe(first(isStable => isStable));
-    const profileLogged = this.orviumService.getProfile().pipe(
+    const profileLogged = this.profileService.getProfile().pipe(
       first(profile => profile !== undefined));
     const notificationsPolling = timer(1000, 60000).pipe(
       switchMap(tick => this.getNotificationsFromApi()),
@@ -42,15 +42,15 @@ export class NotificationService {
     concat(appIsStable, profileLogged, notificationsPolling).subscribe();
   }
 
-  private getNotificationsFromApi(): Observable<AppNotification[]> {
-    return this.http.get<AppNotification[]>(`${environment.apiEndpoint}/notifications/myNotifications`);
-  }
-
   readNotification(notificationId: string): Observable<unknown> {
     return this.http.patch(`${environment.apiEndpoint}/notifications/${notificationId}/read`, null);
   }
 
-  getNotifications(): BehaviorSubject<AppNotification[]> {
+  getNotifications(): BehaviorSubject<AppNotificationDTO[]> {
     return this.notifications;
+  }
+
+  private getNotificationsFromApi(): Observable<AppNotificationDTO[]> {
+    return this.http.get<AppNotificationDTO[]>(`${environment.apiEndpoint}/notifications/myNotifications`);
   }
 }
