@@ -3,7 +3,7 @@ import { CommonModule } from '@angular/common';
 import { ProfileComponent } from './profile/profile.component';
 import { OnboardingComponent } from './onboarding/onboarding.component';
 import { AuthGuardService } from '../services/auth-guard.service';
-import { ProfileResolver } from '../shared/orvium.resolvers';
+import { ProfileResolver, PublicProfileResolver } from '../shared/orvium.resolvers';
 import { RouterModule, Routes } from '@angular/router';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { MatStepperModule } from '@angular/material/stepper';
@@ -19,6 +19,12 @@ import { GravatarModule } from 'ngx-gravatar';
 import { MatButtonModule } from '@angular/material/button';
 import { MatInputModule } from '@angular/material/input';
 import { MatTooltipModule } from '@angular/material/tooltip';
+import { PublicProfileComponent } from './public-profile/public-profile.component';
+import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
+import { ShareModule } from 'ngx-sharebuttons';
+import { SharedModule } from '../shared/shared.module';
+import { MatCheckboxModule } from '@angular/material/checkbox';
+import { OrviumUxLibModule } from '@orvium/ux-components';
 
 const routes: Routes = [
   {
@@ -32,12 +38,17 @@ const routes: Routes = [
     canActivate: [AuthGuardService],
     resolve: { profile: ProfileResolver }
   },
+  {
+    path: ':nickname', component: PublicProfileComponent,
+    resolve: { profile: PublicProfileResolver }
+  },
 ];
 
 @NgModule({
   declarations: [
     ProfileComponent,
     OnboardingComponent,
+    PublicProfileComponent
   ],
   imports: [
     CommonModule,
@@ -52,11 +63,16 @@ const routes: Routes = [
     MatCardModule,
     MatChipsModule,
     MatDatepickerModule,
-    NgCircleProgressModule,
+    NgCircleProgressModule.forRoot({}),
     GravatarModule,
     MatInputModule,
     MatButtonModule,
-    MatTooltipModule
+    MatTooltipModule,
+    FontAwesomeModule,
+    ShareModule,
+    SharedModule,
+    MatCheckboxModule,
+    OrviumUxLibModule,
   ],
   exports: [RouterModule]
 })
