@@ -1,10 +1,8 @@
-import { HttpClientModule } from '@angular/common/http';
 import { MatSnackBarModule } from '@angular/material/snack-bar';
 import { RouterTestingModule } from '@angular/router/testing';
 import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { ProfileComponent } from './profile.component';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { Profile } from '../../model/orvium';
 import { ActivatedRoute } from '@angular/router';
 import { HarnessLoader } from '@angular/cdk/testing';
 import { TestbedHarnessEnvironment } from '@angular/cdk/testing/testbed';
@@ -25,37 +23,34 @@ import { NoopAnimationsModule } from '@angular/platform-browser/animations';
 import { OrviumService } from '../../services/orvium.service';
 import { of } from 'rxjs';
 import { DisciplinesService } from '../../services/disciplines.service';
-import { profileTest } from '../../shared/test-data';
+import { ProfileService } from '../profile.service';
+import { HttpClientTestingModule } from '@angular/common/http/testing';
+import { profilePrivateTest } from '../../shared/test-data';
+import { UserPrivateDTO } from '../../model/api';
 
 
 describe('ProfileComponent', () => {
   let component: ProfileComponent;
   let fixture: ComponentFixture<ProfileComponent>;
   let loader: HarnessLoader;
-  const profile: Profile = { ...profileTest, ...{ isOnboarded: true } };
-  const profile2: Profile = {
-    userId: '123412341234',
-    firstName: 'William',
-    lastName: 'Doe',
-    isReviewer: false,
-    isOnboarded: true,
-    email: 'william@orvium.io',
-    emailConfirmed: false,
-    disciplines: ['Abnormal psychology', 'Acoustics'],
-    userType: 'student',
-    gravatar: '0a2aaae0ac1310d1f8e8e68df45fe7b8',
-    percentageComplete: 0,
-    roles: []
-  };
+  const profile: UserPrivateDTO = { ...profilePrivateTest(), isOnboarded: true };
 
   const routeSnapshot = { snapshot: { data: { profile } } };
 
   beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ProfileComponent],
-      imports: [RouterTestingModule, FormsModule, ReactiveFormsModule,
-        MatChipsModule, MatSnackBarModule, HttpClientModule, MatAutocompleteModule,
-        MatCardModule, MatFormFieldModule, MatDatepickerModule,
+      imports: [
+        RouterTestingModule,
+        FormsModule,
+        ReactiveFormsModule,
+        MatChipsModule,
+        MatSnackBarModule,
+        HttpClientTestingModule,
+        MatAutocompleteModule,
+        MatCardModule,
+        MatFormFieldModule,
+        MatDatepickerModule,
         MatNativeDateModule,
         GravatarModule,
         MatInputModule,
@@ -74,10 +69,12 @@ describe('ProfileComponent', () => {
     component = fixture.componentInstance;
     const orviumService = fixture.debugElement.injector.get(OrviumService);
     spyOn(orviumService, 'getDomain').and.returnValue(of({ emailDomain: 'orvium.io' }));
-    spyOn(orviumService, 'getProfile').and.returnValue(of(profile));
+
+    const profileService = fixture.debugElement.injector.get(ProfileService);
+    spyOn(profileService, 'getProfile').and.returnValue(of(profile));
 
     const disciplinesService = fixture.debugElement.injector.get(DisciplinesService);
-    spyOn(disciplinesService, 'getDisciplines').and.returnValue([]);
+    spyOn(disciplinesService, 'getDisciplines').and.returnValue(of([]));
 
     fixture.detectChanges();
   }));
@@ -93,24 +90,28 @@ describe('ProfileComponent', () => {
   it('should refresh profile', async () => {
     let input = await loader.getHarness(MatInputHarness.with({ placeholder: 'John' }));
     let value = await input.getValue();
-    expect(value).toBe('');
-    const newProfile = { ...profile2 };
+    expect(value).toBe('Willian');
+    const newProfile: UserPrivateDTO = { ...profile, firstName: 'Francis' };
     component.refreshProfile(newProfile);
     input = await loader.getHarness(MatInputHarness.with({ placeholder: 'John' }));
     value = await input.getValue();
-    expect(value).toBe('William');
+    expect(value).toBe('Francis');
   });
 
   it('should remove a discipline', async () => {
-    const save = await loader.getHarness(MatButtonHarness.with({ selector: '#save' }));
+    let save = await loader.getHarness(MatButtonHarness.with({ selector: '#save' }));
     let state = await save.isDisabled();
     expect(state).toBe(true);
-    const newProfile = { ...profile2 };
+    const newProfile: UserPrivateDTO = { ...profile, disciplines: ['Acoustics', 'Computing'] };
     component.refreshProfile(newProfile);
     expect(newProfile.disciplines.length).toBe(2);
     component.removeDiscipline('Acoustics');
     expect(newProfile.disciplines.length).toBe(1);
+    fixture.detectChanges();
     state = await save.isDisabled();
+    console.log('STATE', state);
+    expect(component.profileFormGroup.valid).toBe(true);
+    expect(component.profileFormGroup.pristine).toBe(false);
     expect(state).toBe(false);
   });
 
@@ -118,7 +119,7 @@ describe('ProfileComponent', () => {
     const save = await loader.getHarness(MatButtonHarness.with({ text: 'save Save' }));
     let state = await save.isDisabled();
     expect(state).toBe(true);
-    const newProfile = { ...profile2, ...{ disciplines: ['Abnormal psychology'] } };
+    const newProfile = { ...profile, disciplines: ['Abnormal psychology'] };
     component.refreshProfile(newProfile);
     expect(newProfile.disciplines.length).toBe(1);
     const inputDiscipline = fixture.debugElement.query(By.css('#inputDiscipline'));
@@ -137,10 +138,7 @@ describe('ProfileComponent', () => {
     const save = await loader.getHarness(MatButtonHarness.with({ text: 'save Save' }));
     let state = await save.isDisabled();
     expect(state).toBe(true);
-    const newProfile = { ...profile2 };
-
-    component.refreshProfile(newProfile);
-    const input = await loader.getHarness(MatInputHarness.with({ placeholder: 'John' }));
+    const input = await loader.getHarness(MatInputHarness.with({ selector: '#firstNameInput' }));
     await input.setValue('Jaime');
     state = await save.isDisabled();
     expect(state).toBe(false);
@@ -150,8 +148,6 @@ describe('ProfileComponent', () => {
     const save = await loader.getHarness(MatButtonHarness.with({ text: 'save Save' }));
     let state = await save.isDisabled();
     expect(state).toBe(true);
-    const newProfile = { ...profile2 };
-    component.refreshProfile(newProfile);
     const input = await loader.getHarness(MatInputHarness.with({ placeholder: 'Doe' }));
     await input.setValue('');
     state = await save.isDisabled();
