@@ -4,11 +4,11 @@ import { map, startWith } from 'rxjs/operators';
 import { Observable } from 'rxjs';
 import { MatChipInputEvent } from '@angular/material/chips';
 import { MatAutocomplete, MatAutocompleteSelectedEvent } from '@angular/material/autocomplete';
-import { OrviumService } from '../../services/orvium.service';
 import { ActivatedRoute, Router } from '@angular/router';
-import { Discipline, Profile } from '../../model/orvium';
 import { DisciplinesService } from '../../services/disciplines.service';
 import { MatSliderChange } from '@angular/material/slider';
+import { ProfileService } from '../../profile/profile.service';
+import { DisciplineDTO, UserPrivateDTO } from '../../model/api';
 
 @Component({
   selector: 'app-empty-state-reviews',
@@ -16,27 +16,31 @@ import { MatSliderChange } from '@angular/material/slider';
   styleUrls: ['./empty-state-reviews.component.scss']
 })
 export class EmptyStateReviewsComponent implements OnInit {
-  @ViewChild('disciplineInput') disciplineInput: ElementRef<HTMLInputElement>;
-  @ViewChild('autoCompleteDisciplines') matAutocomplete: MatAutocomplete;
+  @ViewChild('disciplineInput') disciplineInput?: ElementRef<HTMLInputElement>;
+  @ViewChild('autoCompleteDisciplines') matAutocomplete?: MatAutocomplete;
   reviewerFormGroup: FormGroup;
-  filteredDisciplines: Observable<Discipline[]>;
+  filteredDisciplines?: Observable<DisciplineDTO[]>;
   disciplinesControl = new FormControl();
   numberSimultaneousReviews = 1;
-  profile: Profile;
+  profile?: UserPrivateDTO;
+  private disciplines: DisciplineDTO[] = [];
 
   constructor(private formBuilder: FormBuilder,
-              private orviumService: OrviumService,
+              private profileService: ProfileService,
               private disciplinesService: DisciplinesService,
               private router: Router,
               private route: ActivatedRoute) {
-  }
-
-  ngOnInit(): void {
     this.reviewerFormGroup = this.formBuilder.group({
       disciplines: [[]],
       reviewersCode: [null, Validators.required]
     });
+  }
 
+  ngOnInit(): void {
+    this.disciplinesService.getDisciplines().subscribe(
+      disciplines => this.disciplines = disciplines
+    );
+    this.disciplinesControl.setValue([]); // Init control value
     this.filteredDisciplines = this.disciplinesControl.valueChanges.pipe(
       startWith(null),
       map((discipline: string | null) => this.filterDisciplines(discipline || '')));
@@ -55,7 +59,7 @@ export class EmptyStateReviewsComponent implements OnInit {
   }
 
   addDiscipline($event: MatChipInputEvent): void {
-    if (!this.matAutocomplete.isOpen) {
+    if (this.matAutocomplete && !this.matAutocomplete.isOpen) {
       const input = $event.input;
       const value = $event.value;
       const disciplines = this.disciplinesControl.value;
@@ -73,25 +77,14 @@ export class EmptyStateReviewsComponent implements OnInit {
     this.reviewerFormGroup.markAsDirty();
   }
 
-  private filterDisciplines(value: string): Discipline[] {
-    if (value) {
-      const filterValue = value.toLowerCase();
-
-      return this.disciplinesService.getDisciplines().filter(discipline =>
-        discipline.name.toLowerCase().includes(filterValue) &&
-        !this.reviewerFormGroup.get('disciplines')?.value.includes(discipline.name))
-        .slice(0, 50);
-    } else {
-      return this.disciplinesService.getDisciplines().filter(discipline =>
-        !this.reviewerFormGroup.get('disciplines')?.value.includes(discipline.name))
-        .slice(0, 50);
-    }
-  }
-
   selected($event: MatAutocompleteSelectedEvent): void {
-    const disciplines = this.reviewerFormGroup.get('disciplines')?.value;
-    disciplines.push($event.option.viewValue);
-    this.disciplineInput.nativeElement.value = '';
+    const disciplines = this.reviewerFormGroup.controls.disciplines?.value;
+    if (disciplines) {
+      disciplines.push($event.option.viewValue);
+    }
+    if (this.disciplineInput) {
+      this.disciplineInput.nativeElement.value = '';
+    }
     this.disciplinesControl.setValue(null);
   }
 
@@ -104,7 +97,7 @@ export class EmptyStateReviewsComponent implements OnInit {
   }
 
   becomeReviewer(): void {
-    this.orviumService.updateProfile({
+    this.profileService.updateProfile({
       simultaneousReviews: this.numberSimultaneousReviews,
       disciplines: this.reviewerFormGroup.get('disciplines')?.value
     })
@@ -113,10 +106,25 @@ export class EmptyStateReviewsComponent implements OnInit {
       });
   }
 
-  refreshProfile(profile: Profile): void {
+  refreshProfile(profile: UserPrivateDTO): void {
     this.profile = profile;
     this.reviewerFormGroup.patchValue(this.profile);
     this.reviewerFormGroup.markAsPristine();
+  }
+
+  private filterDisciplines(value: string): DisciplineDTO[] {
+    if (value) {
+      const filterValue = value.toLowerCase();
+
+      return this.disciplines.filter(discipline =>
+        discipline.name.toLowerCase().includes(filterValue) &&
+        !this.reviewerFormGroup.controls.disciplines?.value?.includes(discipline.name))
+        .slice(0, 50);
+    } else {
+      return this.disciplines.filter(discipline =>
+        !this.reviewerFormGroup.controls.disciplines?.value?.includes(discipline.name))
+        .slice(0, 50);
+    }
   }
 
 }
