@@ -1,14 +1,24 @@
 import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { MatTabsModule } from '@angular/material/tabs';
-import { InvitationsListComponent } from 'src/app/shared/invitations-list/invitations-list.component';
 
 import { InvitationsPanelComponent } from './invitations-panel.component';
-import { MyreviewsComponent } from '../myreviews/myreviews.component';
 import { RouterTestingModule } from '@angular/router/testing';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { MatTableModule } from '@angular/material/table';
 import { MatDividerModule } from '@angular/material/divider';
+import { Component, Input } from '@angular/core';
+import { InviteDTO, ReviewDTO } from '../../model/api';
+
+@Component({ selector: 'app-invitations-list', template: '' })
+class InvitationsListStubComponent {
+  @Input() invites: InviteDTO[] = [];
+}
+
+@Component({ selector: 'app-myreviews', template: '' })
+class MyreviewsStubComponent {
+  @Input() reviews: ReviewDTO[] = [];
+}
 
 describe('InvitationsPanelComponent', () => {
   let component: InvitationsPanelComponent;
@@ -16,8 +26,19 @@ describe('InvitationsPanelComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [InvitationsPanelComponent, InvitationsListComponent, MyreviewsComponent],
-      imports: [HttpClientTestingModule, MatTabsModule, RouterTestingModule, BrowserAnimationsModule, MatTableModule, MatDividerModule]
+      declarations: [
+        InvitationsPanelComponent,
+        InvitationsListStubComponent,
+        MyreviewsStubComponent
+      ],
+      imports: [
+        HttpClientTestingModule,
+        MatTabsModule,
+        RouterTestingModule,
+        BrowserAnimationsModule,
+        MatTableModule,
+        MatDividerModule
+      ]
     })
       .compileComponents();
   });
