@@ -1,26 +1,17 @@
-import { Component, Input, OnInit } from '@angular/core';
-import { OrviumService } from 'src/app/services/orvium.service';
-import { ActivatedRoute, Router } from '@angular/router';
-import { NgxSpinnerService } from 'ngx-spinner';
-import { PeerReview } from '../../model/orvium';
+import { Component, Input } from '@angular/core';
+import { ReviewDTO } from '../../model/api';
 
 @Component({
   selector: 'app-myreviews',
   templateUrl: './myreviews.component.html',
   styleUrls: ['./myreviews.component.scss']
 })
-export class MyreviewsComponent implements OnInit {
+export class MyreviewsComponent {
 
   displayedColumns = ['publication', 'comments', 'decision', 'action'];
-  @Input() reviews: PeerReview[];
+  @Input() reviews: ReviewDTO[] = [];
 
-  constructor(private orviumService: OrviumService,
-              private route: ActivatedRoute,
-              private router: Router,
-              private spinnerService: NgxSpinnerService) {
+  constructor() {
   }
 
-  ngOnInit(): void {
-
-  }
 }
