@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { Deposit, DEPOSIT_STATUS, PeerReview, Profile, REVIEW_STATUS } from '../model/orvium';
+import { DEPOSIT_STATUS, DepositDTO, REVIEW_STATUS, ReviewDTO, UserPrivateDTO } from '../model/api';
 
 @Injectable({
   providedIn: 'root'
@@ -9,27 +9,47 @@ export class ReviewService {
   constructor() {
   }
 
-  canReadReview(user: Profile | undefined, review: PeerReview): boolean {
-    const hasRights = false;
+  canReadReview(user: UserPrivateDTO | undefined, review: ReviewDTO): boolean {
+    let hasRights = false;
 
     // owner has access
     if (user && review.owner === user.userId) {
-      return true;
+      hasRights = true;
     }
 
     // admin has access
-    if (user && user.roles.includes(`admin`)) {
-      return true;
+    if (user && user.roles.includes('admin')) {
+      hasRights = true;
     }
 
     if (review.status === REVIEW_STATUS.published) {
-      return true;
+      hasRights = true;
     }
 
     return hasRights;
   }
 
-  canUpdateReview(user: Profile | undefined, review: PeerReview): boolean {
+  canUpdateReview(user: UserPrivateDTO | undefined, review: ReviewDTO): boolean {
+    let hasRights = false;
+
+    if (!user) {
+      return false;
+    }
+
+    // If deposit is published then restrict the update
+    if (user && user.userId === review.owner && review.status === REVIEW_STATUS.draft) {
+      hasRights = true;
+    }
+
+    // Admin can update
+    if (user.roles.includes('admin')) {
+      hasRights = true;
+    }
+
+    return hasRights;
+  }
+
+  canDeleteReview(user: UserPrivateDTO | undefined, review: ReviewDTO): boolean {
     let hasRights = false;
 
     if (!user) {
@@ -49,27 +69,7 @@ export class ReviewService {
     return hasRights;
   }
 
-  canDeleteReview(user: Profile | undefined, review: PeerReview): boolean {
-    let hasRights = false;
-
-    if (!user) {
-      return false;
-    }
-
-    // If deposit is published then restrict the update
-    if (user.userId === review.owner && review.status === REVIEW_STATUS.draft) {
-      hasRights = true;
-    }
-
-    // Admin can update
-    if (user.roles.includes('admin')) {
-      hasRights = true;
-    }
-
-    return hasRights;
-  }
-
-  canCreateReview(user: Profile | undefined, deposit: Deposit): boolean {
+  canCreateReview(user: UserPrivateDTO | undefined, deposit: DepositDTO): boolean {
     if (!user) {
       return false;
     }
