@@ -1,18 +1,20 @@
 import { Component, OnInit } from '@angular/core';
-import { Deposit, OrviumFile, PeerReview, Profile, REVIEW_DECISION_LOV, REVIEW_STATUS, REVIEW_TYPE, } from 'src/app/model/orvium';
+import { REVIEW_DECISION_LOV } from 'src/app/model/orvium';
 import { ActivatedRoute, Router } from '@angular/router';
 import { OrviumService } from '../../services/orvium.service';
 import { environment } from '../../../environments/environment';
 import { EthereumService } from '../../blockchain/ethereum.service';
-import { MatSnackBar } from '@angular/material/snack-bar';
 import { finalize } from 'rxjs/operators';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { Title } from '@angular/platform-browser';
-import { NgxSpinnerService } from 'ngx-spinner';
 import { NGXLogger } from 'ngx-logger';
 import { BlockchainService } from '../../blockchain/blockchain.service';
 import { canOpenOverleaf } from '../../shared/shared-functions';
 import { HttpResponse } from '@angular/common/http';
+import { ProfileService } from '../../profile/profile.service';
+import { AppSnackBarService } from '../../services/app-snack-bar.service';
+import { DepositDTO, FileMetadata, REVIEW_STATUS, REVIEW_TYPE, ReviewDTO, UserPrivateDTO } from 'src/app/model/api';
+import { OrvSpinnerService } from '@orvium/ux-components';
 
 @Component({
   selector: 'app-reviews-create',
@@ -21,11 +23,11 @@ import { HttpResponse } from '@angular/common/http';
 })
 export class ReviewsCreateComponent implements OnInit {
 
-  peerReview: PeerReview;
-  deposit: Deposit;
-  profile: Profile;
-  files: OrviumFile[] = [];
-  filelink: string;
+  peerReview: ReviewDTO;
+  deposit: DepositDTO;
+  profile?: UserPrivateDTO;
+  files: FileMetadata[] = [];
+  filelink = '';
   environment = environment;
   isOwner = false;
   reviewForm: FormGroup;
@@ -37,34 +39,35 @@ export class ReviewsCreateComponent implements OnInit {
 
   constructor(
     private orviumService: OrviumService,
+    private profileService: ProfileService,
     public blockchainService: BlockchainService,
     private route: ActivatedRoute,
-    private spinnerService: NgxSpinnerService,
-    private snackBar: MatSnackBar,
+    private spinnerService: OrvSpinnerService,
+    private snackBar: AppSnackBarService,
     public ethereumService: EthereumService,
     private router: Router,
     private titleService: Title,
     private formBuilder: FormBuilder,
-    private logger: NGXLogger) {
-  }
-
-  ngOnInit(): void {
-    this.peerReview = this.route.snapshot.data.review;
-    this.deposit = this.peerReview.deposit;
-    const profile = this.orviumService.profile.getValue();
-
-    if (!profile) {
-      throw new Error('Profile is undefined');
-    }
-
-    this.profile = profile;
-
+    private logger: NGXLogger
+  ) {
     this.reviewForm = this.formBuilder.group({
       author: [null, [Validators.required, Validators.pattern(/.*\S.*/)]],
       comments: [],
       decision: [],
       revealReviewerIdentity: [],
     });
+
+    this.peerReview = this.route.snapshot.data.review;
+    this.deposit = this.peerReview.deposit;
+  }
+
+  ngOnInit(): void {
+    const profile = this.profileService.profile.getValue();
+
+    if (!profile) {
+      throw new Error('Profile is undefined');
+    }
+    this.profile = profile;
 
     this.refreshPeerReview(this.peerReview);
 
@@ -90,17 +93,17 @@ export class ReviewsCreateComponent implements OnInit {
       .subscribe(response => {
         this.refreshPeerReview(response);
         this.reviewForm.markAsPristine();
-        this.snackBar.open('Peer Review saved', 'Dismiss', { panelClass: ['ok-snackbar'] });
+        this.snackBar.info('Peer Review saved');
       });
   }
 
   canPublish(): boolean {
     if (!this.peerReview.file) {
-      this.snackBar.open('Review report file is missing', 'Dismiss', { panelClass: ['error-snackbar'] });
+      this.snackBar.error('Review report file is missing');
       return false;
     }
     if (!this.reviewForm.pristine) {
-      this.snackBar.open('Save all your changes first', 'Dismiss', { panelClass: ['error-snackbar'] });
+      this.snackBar.error('Save all your changes first');
       return false;
     }
     ['author', 'comments', 'decision']
@@ -119,7 +122,7 @@ export class ReviewsCreateComponent implements OnInit {
         status: REVIEW_STATUS.published
       }).subscribe(response => {
         this.refreshPeerReview(response);
-        this.snackBar.open('Peer Review Published', 'Dismiss', { panelClass: ['ok-snackbar'] });
+        this.snackBar.info('Peer Review Published');
       });
     }
   }
@@ -132,18 +135,29 @@ export class ReviewsCreateComponent implements OnInit {
       })
     )
       .subscribe(response => {
-        this.snackBar.open('Review deleted', 'Dismiss', { panelClass: ['ok-snackbar'] });
+        this.snackBar.info('Review deleted');
         return this.router.navigate(['/deposits/' + this.deposit._id + '/view']);
       });
   }
 
-  proofOwnership(): void {
-    if (!this.peerReview.file?.keccak256) {
-      this.snackBar.open('Upload a file first to publish to blockchain', 'Dismiss', { panelClass: ['error-snackbar'] });
+  async proofOwnership(): Promise<void> {
+    if (!this.peerReview.file) {
+      this.snackBar.error('Upload a file first');
       return;
     }
 
-    this.ethereumService.publicationProofOwnership(this.peerReview.file.keccak256).subscribe(transaction => {
+    // Calculate hash
+    const fileUrl = `${environment.apiEndpoint}/reviews/${this.peerReview._id}/file`;
+    let responseFile = await fetch(fileUrl);
+    let arrayBuffer = await responseFile.arrayBuffer();
+    const fileHash = this.ethereumService.hashFile(arrayBuffer);
+
+    if (!fileHash) {
+      this.snackBar.error('Incorrect peer review hash');
+      return;
+    }
+
+    this.ethereumService.publicationProofOwnership(fileHash).subscribe(transaction => {
       const spinnerName = 'spinnerProof';
       this.logger.debug('Transaction', transaction);
       this.spinnerService.show(spinnerName);
@@ -167,13 +181,31 @@ export class ReviewsCreateComponent implements OnInit {
         }
         if (this.ethereumService.currentNetwork.value) {
           this.peerReview.transactions[this.ethereumService.currentNetwork.value.name] = receipt;
+          this.peerReview.keccak256 = fileHash;
         }
         this.save();
       });
     });
   }
 
-  private refreshPeerReview(peerReview: PeerReview): void {
+  async filesUploaded($event: { originalEvent: HttpResponse<unknown>, files: File[] }): Promise<void> {
+    const updatedPeerReview = await this.orviumService.getPeerReview(this.peerReview._id).toPromise();
+    this.refreshPeerReview(updatedPeerReview);
+  }
+
+  beforeUpload($event: unknown): boolean {
+    if (!this.reviewForm.pristine) {
+      if (this.reviewForm.valid) {
+        this.save();
+      } else {
+        this.snackBar.error('Some data is invalid, please review');
+        return false;
+      }
+    }
+    return true;
+  }
+
+  private refreshPeerReview(peerReview: ReviewDTO): void {
     this.titleService.setTitle(`Peer Review by ${peerReview.author}`);
     this.peerReview = peerReview;
     this.reviewForm.reset(this.peerReview);
@@ -187,19 +219,5 @@ export class ReviewsCreateComponent implements OnInit {
     if (this.peerReview.status === REVIEW_STATUS.published) {
       this.reviewForm?.disable();
     }
-  }
-
-  filesUploaded($event: { originalEvent: HttpResponse<unknown>, files: File[] }): void {
-    const updatedPeerReview = $event.originalEvent.body as PeerReview;
-    this.refreshPeerReview(updatedPeerReview);
-  }
-
-  onSelectFiles($event: unknown): boolean {
-    if (!this.reviewForm.pristine) {
-      this.snackBar.open('You need to save all your changes before uploading files',
-        'Dismiss', { panelClass: ['error-snackbar'] });
-      return false;
-    }
-    return true;
   }
 }
