@@ -1,49 +1,23 @@
-import {
-  AfterViewInit,
-  ChangeDetectionStrategy,
-  Component,
-  ElementRef,
-  EventEmitter,
-  Input,
-  NgZone,
-  OnDestroy,
-  Output,
-  Renderer2,
-  ViewChild
-} from '@angular/core';
-import { DomSanitizer, SafeUrl } from '@angular/platform-browser';
-import { HttpClient, HttpEvent, HttpEventType, HttpHeaders, HttpResponse } from '@angular/common/http';
-import { MatSnackBar } from '@angular/material/snack-bar';
-
-interface FileExtended extends File {
-  objectURL: SafeUrl;
-}
+import { Component, ElementRef, EventEmitter, Input, OnChanges, OnInit, Output, ViewChild } from '@angular/core';
+import { HttpClient, HttpErrorResponse, HttpEvent, HttpEventType, HttpResponse } from '@angular/common/http';
+import { AppSnackBarService } from '../../services/app-snack-bar.service';
 
 @Component({
   selector: 'app-fileupload',
   templateUrl: './fileupload.component.html',
   styleUrls: ['./fileupload.component.scss'],
-  changeDetection: ChangeDetectionStrategy.Default
 })
-export class FileuploadComponent implements AfterViewInit, OnDestroy {
+export class FileuploadComponent {
 
-  @Input() name: string;
+  @Input() name = 'file';
 
-  @Input() url: string;
+  @Input() url?: string;
 
-  @Input() method = 'POST';
+  @Input() accept?: string;
 
-  @Input() multiple: boolean;
+  @Input() disabled = false;
 
-  @Input() accept: string;
-
-  @Input() disabled: boolean;
-
-  @Input() auto: boolean;
-
-  @Input() withCredentials: boolean;
-
-  @Input() maxFileSize: number;
+  @Input() maxFileSize?: number;
 
   @Input() invalidFileSizeMessageSummary = '{0}: Invalid file size, ';
 
@@ -57,29 +31,17 @@ export class FileuploadComponent implements AfterViewInit, OnDestroy {
 
   @Input() invalidFileLimitMessageSummary = 'Maximum number of files exceeded, ';
 
-  @Input() previewWidth = 50;
-
   @Input() chooseLabel = 'Choose';
 
   @Input() uploadLabel = 'Upload';
 
   @Input() cancelLabel = 'Cancel';
 
-  @Input() chooseIcon = 'pi pi-plus';
+  @Input() customUpload = false;
 
-  @Input() uploadIcon = 'pi pi-upload';
+  @Input() fileLimit = 0;
 
-  @Input() cancelIcon = 'pi pi-times';
-
-  @Input() showUploadButton = true;
-
-  @Input() showCancelButton = true;
-
-  @Input() headers: HttpHeaders;
-
-  @Input() customUpload: boolean;
-
-  @Input() fileLimit: number;
+  @Input() uploadedFileCount = 0;
 
   @Output() beforeUpload: EventEmitter<unknown> = new EventEmitter();
 
@@ -91,98 +53,59 @@ export class FileuploadComponent implements AfterViewInit, OnDestroy {
 
   @Output() clear: EventEmitter<void> = new EventEmitter();
 
-  @Output() remove: EventEmitter<unknown> = new EventEmitter();
-
   @Output() fileSelect: EventEmitter<unknown> = new EventEmitter();
 
   @Output() fileProgress: EventEmitter<unknown> = new EventEmitter();
 
   @Output() uploadHandler: EventEmitter<{ files: File[] }> = new EventEmitter();
 
-  @ViewChild('advancedfileinput') advancedFileInput: ElementRef;
+  @ViewChild('advancedfileinput') advancedFileInput?: ElementRef;
 
-  @ViewChild('basicfileinput') basicFileInput: ElementRef;
+  @ViewChild('content') content?: ElementRef;
+  public progress = 0;
+  uploading = false;
 
-  @ViewChild('content') content: ElementRef;
+  constructor(
+    private http: HttpClient,
+    private snackBar: AppSnackBarService
+  ) {
+  }
+
+  public _files: File[] = [];
+
+  get files(): File[] {
+    return this._files;
+  }
 
   @Input() set files(files) {
     this._files = [];
 
     for (const file of files) {
       if (this.validate(file)) {
-        if (this.isImage(file)) {
-          (file as FileExtended).objectURL = this.sanitizer.bypassSecurityTrustUrl((window.URL.createObjectURL(file)));
-        }
-
         this._files.push(file);
       }
     }
   }
 
-  get files(): File[] {
-    return this._files;
-  }
-
-  public _files: File[] = [];
-
-  public progress = 0;
-
-  public dragHighlight: boolean;
-
-  public uploadedFileCount = 0;
-
-  focus: boolean;
-
-  uploading: boolean;
-
-  constructor(private el: ElementRef,
-              public sanitizer: DomSanitizer,
-              public zone: NgZone,
-              private http: HttpClient,
-              private snackBar: MatSnackBar,
-              private renderer: Renderer2) {
-  }
-
-  ngAfterViewInit(): void {
-    this.zone.runOutsideAngular(() => {
-      if (this.content) {
-        this.content.nativeElement.addEventListener('dragover', this.onDragOver.bind(this));
-      }
-    });
-  }
-
   onFileSelect(event: Event): void {
-    if (!this.multiple) {
-      this.files = [];
-    }
-
-    const files = (event instanceof DragEvent && event.dataTransfer) ?
-      event.dataTransfer.files : (event.target as HTMLInputElement).files;
+    const files = (event.target as HTMLInputElement).files;
 
     for (let i = 0; files && i < files.length; i++) {
       const file = files[i];
       if (!this.isFileSelected(file)) {
         if (this.validate(file)) {
-          if (this.isImage(file)) {
-            (file as FileExtended).objectURL = this.sanitizer.bypassSecurityTrustUrl((window.URL.createObjectURL(file)));
-          }
           this.files.push(file);
         }
       }
     }
-
     this.fileSelect.emit({ originalEvent: event, files, currentFiles: this.files });
 
-    if (this.fileLimit) {
+    if (this.fileLimit > 0) {
       this.checkFileLimit();
     }
 
-    if (this.hasFiles() && this.auto && !this.isFileLimitExceeded()) {
+    if (this.hasFiles() && !this.isFileLimitExceeded()) {
       this.upload();
-    }
-
-    if (event.type === 'drop') {
-      this.clearInputElement();
     }
   }
 
@@ -199,13 +122,13 @@ export class FileuploadComponent implements AfterViewInit, OnDestroy {
   validate(file: File): boolean {
     if (this.accept && !this.isFileTypeValid(file)) {
       const errorMessage = this.invalidFileTypeMessageDetail.replace('{0}', this.accept);
-      this.snackBar.open(errorMessage, 'Dismiss', { panelClass: ['error-snackbar'] });
+      this.snackBar.error(errorMessage);
       return false;
     }
 
     if (this.maxFileSize && file.size > this.maxFileSize) {
       const errorMessage = this.invalidFileSizeMessageDetail.replace('{0}', this.formatSize(this.maxFileSize));
-      this.snackBar.open(errorMessage, 'Dismiss', { panelClass: ['error-snackbar'] });
+      this.snackBar.error(errorMessage);
       return false;
     }
 
@@ -213,13 +136,16 @@ export class FileuploadComponent implements AfterViewInit, OnDestroy {
   }
 
   isFileTypeValid(file: File): boolean {
-    const acceptableTypes = this.accept.split(',').map(type => type.trim());
-    for (const type of acceptableTypes) {
-      const acceptable = this.isWildcard(type) ? this.getTypeClass(file.type) === this.getTypeClass(type)
-        : file.type === type || this.getFileExtension(file).toLowerCase() === type.toLowerCase();
+    if (this.accept) {
+      const acceptableTypes = this.accept.split(',').map(type => type.trim());
+      for (const type of acceptableTypes) {
+        const acceptable = this.isWildcard(type)
+          ? this.getTypeClass(file.type) === this.getTypeClass(type)
+          : file.type === type || this.getFileExtension(file).toLowerCase() === type.toLowerCase();
 
-      if (acceptable) {
-        return true;
+        if (acceptable) {
+          return true;
+        }
       }
     }
 
@@ -238,17 +164,10 @@ export class FileuploadComponent implements AfterViewInit, OnDestroy {
     return '.' + file.name.split('.').pop();
   }
 
-  isImage(file: File): boolean {
-    return /^image\//.test(file.type);
-  }
+  async upload(): Promise<void> {
 
-  onImageLoad(img: HTMLImageElement): void {
-    window.URL.revokeObjectURL(img.src);
-  }
-
-  upload(): void {
     if (this.customUpload) {
-      if (this.fileLimit) {
+      if (this.fileLimit>0) {
         this.uploadedFileCount += this.files.length;
       }
 
@@ -267,9 +186,25 @@ export class FileuploadComponent implements AfterViewInit, OnDestroy {
         formData.append(this.name, file, file.name);
       }
 
-      this.http.post(this.url, formData, {
-        headers: this.headers, reportProgress: true, observe: 'events', withCredentials: this.withCredentials
-      }).subscribe((event: HttpEvent<unknown>) => {
+      const fileObject = this.files[0];
+
+      const file = {
+        name: fileObject.name,
+        type: fileObject.type,
+        size: fileObject.size,
+        lastModified: fileObject.lastModified
+      };
+
+      if (!this.url) {
+        throw new Error('No presigned url configured');
+      }
+
+      let res = await this.http.post<{ signedUrl: string }>(this.url, { file }).toPromise();
+
+      this.http.put(res.signedUrl, fileObject, {
+        reportProgress: true, observe: 'events'
+      }).subscribe(
+        (event: HttpEvent<unknown>) => {
           switch (event.type) {
             case HttpEventType.Sent:
               this.send.emit({
@@ -282,13 +217,14 @@ export class FileuploadComponent implements AfterViewInit, OnDestroy {
               this.progress = 0;
 
               if (event.status >= 200 && event.status < 300) {
-                if (this.fileLimit) {
+                if (this.fileLimit>0) {
                   this.uploadedFileCount += this.files.length;
                 }
 
                 this.fileUpload.emit({ originalEvent: event, files: this.files });
               } else {
                 this.fileError.emit({ files: this.files });
+                this.snackBar.error(event.statusText);
               }
 
               this.onClear();
@@ -303,10 +239,12 @@ export class FileuploadComponent implements AfterViewInit, OnDestroy {
             }
           }
         },
-        error => {
+        (error: HttpErrorResponse) => {
           this.uploading = false;
           this.fileError.emit({ files: this.files, error });
-        });
+          this.snackBar.error(error.message);
+        }
+      );
     }
   }
 
@@ -316,17 +254,7 @@ export class FileuploadComponent implements AfterViewInit, OnDestroy {
     this.clearInputElement();
   }
 
-  onRemove(event: Event, index: number): void {
-    this.clearInputElement();
-    this.remove.emit({ originalEvent: event, file: this.files[index] });
-    this.files.splice(index, 1);
-  }
-
   isFileLimitExceeded(): boolean {
-    if (this.fileLimit && this.fileLimit <= this.files.length + this.uploadedFileCount && this.focus) {
-      this.focus = false;
-    }
-
     return (this.fileLimit > 0) && this.fileLimit < this.files.length + this.uploadedFileCount;
   }
 
@@ -337,7 +265,8 @@ export class FileuploadComponent implements AfterViewInit, OnDestroy {
   checkFileLimit(): void {
     if (this.isFileLimitExceeded()) {
       const errorMessage = this.invalidFileLimitMessageDetail.replace('{0}', this.fileLimit.toString());
-      this.snackBar.open(errorMessage, 'Dismiss', { panelClass: ['error-snackbar'] });
+      this.snackBar.error(errorMessage);
+      this.onClear();
     }
   }
 
@@ -345,61 +274,10 @@ export class FileuploadComponent implements AfterViewInit, OnDestroy {
     if (this.advancedFileInput && this.advancedFileInput.nativeElement) {
       this.advancedFileInput.nativeElement.value = '';
     }
-
-    if (this.basicFileInput && this.basicFileInput.nativeElement) {
-      this.basicFileInput.nativeElement.value = '';
-    }
   }
 
   hasFiles(): boolean {
     return this.files && this.files.length > 0;
-  }
-
-  onDragEnter(e: DragEvent): void {
-    if (!this.disabled) {
-      e.stopPropagation();
-      e.preventDefault();
-    }
-  }
-
-  onDragOver(e: DragEvent): void {
-    if (!this.disabled) {
-      this.renderer.addClass(this.content.nativeElement, 'ui-fileupload-highlight');
-      this.dragHighlight = true;
-      e.stopPropagation();
-      e.preventDefault();
-    }
-  }
-
-  onDragLeave(event: DragEvent): void {
-    if (!this.disabled) {
-      this.renderer.removeClass(this.content.nativeElement, 'ui-fileupload-highlight');
-    }
-  }
-
-  onDrop(event: DragEvent): void {
-    if (!this.disabled) {
-      this.renderer.removeClass(this.content.nativeElement, 'ui-fileupload-highlight');
-      event.stopPropagation();
-      event.preventDefault();
-
-      const files = (event instanceof DragEvent && event.dataTransfer) ?
-        event.dataTransfer.files : (event.target as HTMLInputElement).files;
-
-      const allowDrop = this.multiple || (files && files.length === 1);
-
-      if (allowDrop) {
-        this.onFileSelect(event);
-      }
-    }
-  }
-
-  onFocus(): void {
-    this.focus = true;
-  }
-
-  onBlur(): void {
-    this.focus = false;
   }
 
   formatSize(bytes: number): string {
@@ -412,11 +290,5 @@ export class FileuploadComponent implements AfterViewInit, OnDestroy {
     const i = Math.floor(Math.log(bytes) / Math.log(k));
 
     return parseFloat((bytes / Math.pow(k, i)).toFixed(dm)) + ' ' + sizes[i];
-  }
-
-  ngOnDestroy(): void {
-    if (this.content && this.content.nativeElement) {
-      this.content.nativeElement.removeEventListener('dragover', this.onDragOver);
-    }
   }
 }
