@@ -1,14 +1,14 @@
 import { Component, Inject, OnInit, PLATFORM_ID, ViewChild } from '@angular/core';
 import { MatSidenav } from '@angular/material/sidenav';
 import { SidenavService } from '../services/sidenav.service';
-import { OrviumService } from '../services/orvium.service';
-import { MatSnackBar } from '@angular/material/snack-bar';
-import { Profile } from '../model/orvium';
 import { Router } from '@angular/router';
 import { BreakpointObserver } from '@angular/cdk/layout';
 import { InviteComponent } from '../invite/invite.component';
 import { MatDialog } from '@angular/material/dialog';
 import { DOCUMENT } from '@angular/common';
+import { ProfileService } from '../profile/profile.service';
+import { AppSnackBarService } from '../services/app-snack-bar.service';
+import { UserPrivateDTO } from '../model/api';
 
 @Component({
   selector: 'app-side-nav',
@@ -16,31 +16,36 @@ import { DOCUMENT } from '@angular/common';
   styleUrls: ['./side-nav.component.scss']
 })
 export class SideNavComponent implements OnInit {
-  @ViewChild('sidenav', { static: true }) sidenav: MatSidenav;
-  profile: Profile;
+  @ViewChild('sidenav', { static: true }) sidenav?: MatSidenav;
+  @ViewChild('rightSidenav', { static: true }) rightSidenav?: MatSidenav;
+  profile?: UserPrivateDTO;
   smallScreen = false;
 
-  constructor(private sidenavService: SidenavService,
-              private snackBar: MatSnackBar,
-              public router: Router,
-              public orviumService: OrviumService,
-              public breakpointObserver: BreakpointObserver,
-              private dialog: MatDialog,
-              @Inject(PLATFORM_ID) private platformId: string,
-              @Inject(DOCUMENT) private document: Document) {
+  constructor(
+    private sidenavService: SidenavService,
+    private snackBar: AppSnackBarService,
+    public router: Router,
+    private profileService: ProfileService,
+    public breakpointObserver: BreakpointObserver,
+    private dialog: MatDialog,
+    @Inject(PLATFORM_ID) private platformId: string,
+    @Inject(DOCUMENT) private document: Document) {
   }
 
-
   ngOnInit(): void {
-    this.sidenavService.setSidenav(this.sidenav);
-    this.orviumService.getProfile().subscribe(profileRefreshed => {
+    if (this.sidenav) {
+      this.sidenavService.setSidenav(this.sidenav);
+    }
+    if (this.rightSidenav) {
+      this.sidenavService.setRightSidenav(this.rightSidenav);
+    }
+    this.profileService.getProfile().subscribe(profileRefreshed => {
       if (profileRefreshed) {
         this.profile = profileRefreshed;
       }
     });
     this.breakpointObserver.observe('(max-width: 768px)')
       .subscribe(result => {
-        console.log('screen', result);
         this.smallScreen = result.matches;
       });
 
@@ -50,8 +55,8 @@ export class SideNavComponent implements OnInit {
   }
 
   isReviewer(): void {
-    if (!this.profile?.isOnboarded) {
-      this.snackBar.open('Complete your profile first to become a reviewer', 'Dismiss', { panelClass: ['info-snackbar'] });
+    if (!this.profile?.isOnboarded || !this.profile?.emailConfirmed) {
+      this.snackBar.info('Complete your profile & confirm your email first to become a reviewer');
     } else if (!this.profile?.isReviewer) {
       this.router.navigateByUrl('/reviews');
     }
@@ -61,17 +66,21 @@ export class SideNavComponent implements OnInit {
     this.sidenavService.toggle();
   }
 
+  toggleRightSidenav(): void {
+    this.sidenavService.toggleRight();
+  }
+
   closeSidenav(): void {
     if (this.smallScreen) {
       this.sidenavService.close();
+      this.sidenavService.closeRight();
     }
+
   }
 
   openInviteDialog(): void {
-    if (!this.profile.isOnboarded) {
-      this.snackBar.open('You need to complete your profile first',
-        'Dismiss',
-        { panelClass: ['info-snackbar'] });
+    if (this.profile && !this.profile.isOnboarded) {
+      this.snackBar.info('You need to complete your profile first');
       return;
     }
 
