@@ -1,5 +1,4 @@
 import { MatDialogModule } from '@angular/material/dialog';
-import { HttpClientModule } from '@angular/common/http';
 import { MatSnackBarModule } from '@angular/material/snack-bar';
 import { MatMenuModule } from '@angular/material/menu';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
@@ -8,7 +7,6 @@ import { ToolbarComponent } from './toolbar.component';
 import { RouterTestingModule } from '@angular/router/testing';
 import { EthereumService } from '../blockchain/ethereum.service';
 import { Router } from '@angular/router';
-import { OrviumService } from '../services/orvium.service';
 import { of } from 'rxjs';
 import { SidenavService } from '../services/sidenav.service';
 import { NgxSmartModalModule } from 'ngx-smart-modal';
@@ -22,42 +20,41 @@ import { NoopAnimationsModule } from '@angular/platform-browser/animations';
 import { MatBadgeModule } from '@angular/material/badge';
 import { GravatarModule } from 'ngx-gravatar';
 import { MatIconTestingModule } from '@angular/material/icon/testing';
-import { MatIconModule } from '@angular/material/icon';
 import { NotificationService } from '../notification/notification.service';
-import SpyObj = jasmine.SpyObj;
+import { ProfileService } from '../profile/profile.service';
+import { HttpClientTestingModule } from '@angular/common/http/testing';
+import { MatIconModule } from '@angular/material/icon';
 
 
 describe('ToolbarComponent', () => {
   let component: ToolbarComponent;
   let fixture: ComponentFixture<ToolbarComponent>;
-  let ethServiceSpy: jasmine.SpyObj<EthereumService>;
   let router: Router;
 
   beforeEach(waitForAsync(() => {
-    const ethSpy = jasmine.createSpyObj('EthereumService', ['getNetwork', 'isAvailable']);
-    const orviumSpy = jasmine.createSpyObj('OrviumService', ['getProfile']);
+    const ethSpy = jasmine.createSpyObj('EthereumService', ['getNetwork']);
+    const profileServiceSpy = jasmine.createSpyObj('ProfileService', ['getProfile']);
     const sidenavSpy = jasmine.createSpyObj('SidenavService', ['toggle']);
 
     const notificationSpy = jasmine.createSpyObj('NotificationService', ['getNotifications']);
     notificationSpy.getNotifications.and.returnValue(of([]));
 
-    ethSpy.isAvailable.and.returnValue(true);
     ethSpy.getNetwork.and.returnValue(Promise.resolve('ropsten'));
-    orviumSpy.getProfile.and.returnValue(of({
+    profileServiceSpy.getProfile.and.returnValue(of({
       gravatar: '0a2aaae0ac1310d1f8e8e68df45fe7b8'
     }));
     sidenavSpy.toggle.and.returnValue(true);
 
     TestBed.configureTestingModule({
       declarations: [ToolbarComponent],
-      imports: [FormsModule, RouterTestingModule.withRoutes([]),
+      imports: [
         FormsModule,
+        RouterTestingModule.withRoutes([]),
         ReactiveFormsModule,
         MatMenuModule,
-        MatSnackBarModule,
         MatToolbarModule,
-        MatFormFieldModule,
         MatIconModule,
+        MatFormFieldModule,
         MatIconTestingModule,
         MatSlideToggleModule,
         MatDividerModule,
@@ -65,13 +62,15 @@ describe('ToolbarComponent', () => {
         MatBadgeModule,
         GravatarModule,
         NoopAnimationsModule,
-        HttpClientModule,
+        HttpClientTestingModule,
         MatListModule,
         MatDialogModule,
-        NgxSmartModalModule.forRoot()],
+        NgxSmartModalModule.forRoot(),
+        MatSnackBarModule,
+      ],
       providers: [
         { provide: EthereumService, useValue: ethSpy },
-        { provide: OrviumService, useValue: orviumSpy },
+        { provide: ProfileService, useValue: profileServiceSpy },
         { provide: SidenavService, useValue: sidenavSpy },
         {
           provide: NotificationService,
@@ -80,8 +79,6 @@ describe('ToolbarComponent', () => {
       ]
     }).compileComponents();
     router = TestBed.inject(Router);
-    ethServiceSpy = TestBed.inject(EthereumService) as SpyObj<EthereumService>;
-    ethServiceSpy.isAvailable.and.returnValue(true);
   }));
 
   beforeEach(() => {
